Cleanup

+++ Luci-Engine/src/Luci/Renderer/Buffer.h
@@ -101,7 +101,7 @@ namespace Luci {
 		virtual void SetLayout(const BufferLayout& layout) = 0;
 		virtual const BufferLayout& GetLayout() const = 0;
 
-		virtual uint32_t GetCount() const = 0;
+		virtual size_t GetCount() const = 0;
 	};
 
 	class IndexBuffer {
@@ -112,7 +112,7 @@ namespace Luci {
 		virtual void Bind() const = 0;
 		virtual void UnBind() const = 0;
 
-		virtual uint32_t GetCount() const = 0;
+		virtual size_t GetCount() const = 0;
 	};
 
 }
