+++ Luci-Engine/src/Platform/OpenGL/OpenGLBuffer.h
@@ -15,11 +15,11 @@ namespace Luci {
 		virtual void SetLayout(const BufferLayout& layout) override { m_Layout = layout; }
 		virtual const BufferLayout& GetLayout() const override { return m_Layout; }
 
-		virtual uint32_t GetCount() const { return m_Count; }
+		virtual size_t GetCount() const { return m_Count; }
 
 	private:
 		uint32_t m_RendererID;
-		uint32_t m_Count;
+		size_t m_Count;
 		BufferLayout m_Layout;
 	};
 
@@ -31,11 +31,11 @@ namespace Luci {
 		virtual void Bind() const override;
 		virtual void UnBind() const override;
 
-		virtual uint32_t GetCount() const { return m_Count; }
+		virtual size_t GetCount() const { return m_Count; }
 
 	private:
 		uint32_t m_RendererID;
-		uint32_t m_Count;
+		size_t m_Count;
 	};
 
 }
